# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/internal/admin/refunds/{ticketId}/review" : {
      "post" : {
        "tags" : [ "admin-refund-controller" ],
        "operationId" : "addReviewNotes",
        "parameters" : [ {
          "name" : "ticketId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SupportTicket"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/refunds/{ticketId}/resolve" : {
      "post" : {
        "tags" : [ "admin-refund-controller" ],
        "operationId" : "resolveTicket",
        "parameters" : [ {
          "name" : "ticketId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-User-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SupportTicket"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/orders/{orderId}/participants" : {
      "get" : {
        "tags" : [ "internal-order-controller" ],
        "operationId" : "getOrderParticipants",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/orders/driver/{driverId}/history" : {
      "get" : {
        "tags" : [ "internal-order-controller" ],
        "operationId" : "getOrderHistoryForDriver",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "date",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageOrder"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/orders/driver/{driverId}/active" : {
      "get" : {
        "tags" : [ "internal-order-controller" ],
        "operationId" : "getActiveOrdersForDriver",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageOrder"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/refunds" : {
      "get" : {
        "tags" : [ "admin-refund-controller" ],
        "operationId" : "getTickets",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageSupportTicket"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/orders/unassigned" : {
      "get" : {
        "tags" : [ "admin-order-controller" ],
        "operationId" : "getUnassignedOrders_1",
        "parameters" : [ {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageOrder"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ReviewRequest" : {
        "type" : "object",
        "properties" : {
          "notes" : {
            "type" : "string"
          }
        }
      },
      "SupportTicket" : {
        "required" : [ "createdAt", "customerId", "id", "orderId", "reason", "status" ],
        "type" : "object",
        "properties" : {
          "version" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reason" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "OPEN", "IN_REVIEW", "RESOLVED", "REJECTED" ]
          },
          "resolutionNotes" : {
            "type" : "string"
          },
          "resolvedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "resolvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "chatSessionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "requestedRefundItems" : {
            "type" : "string"
          },
          "refundAmount" : {
            "type" : "number"
          },
          "restaurantComments" : {
            "type" : "string"
          },
          "riderComments" : {
            "type" : "string"
          }
        }
      },
      "ResolveRequest" : {
        "required" : [ "approved" ],
        "type" : "object",
        "properties" : {
          "approved" : {
            "type" : "boolean"
          },
          "notes" : {
            "type" : "string"
          },
          "faultType" : {
            "type" : "string"
          },
          "overrideAmount" : {
            "type" : "number"
          }
        }
      },
...
      "PageableObject" : {
        "required" : [ "offset", "pageNumber", "pageSize", "paged", "unpaged" ],
        "type" : "object",
        "properties" : {
          "unpaged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "paged" : {
            "type" : "boolean"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "SortObject" : {
        "required" : [ "empty", "sorted", "unsorted" ],
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
...
      "Order" : {
        "required" : [ "createdAt", "customerId", "id", "restaurantId", "status", "totalAmount", "updatedAt" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerName" : {
            "type" : "string"
          },
          "restaurantId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "restaurantName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "CREATED", "PENDING_ACCEPTANCE", "AWAITING_DELAY_APPROVAL", "ACCEPTED", "PREPARING", "READY_FOR_PICKUP", "HANDED_OVER", "CANCELLED", "CANCELLED_BY_RESTAURANT" ]
          },
          "paymentMethod" : {
            "type" : "string",
            "enum" : [ "CARD", "UPI", "WALLET", "COD" ]
          },
          "cashCollectedAmount" : {
            "type" : "number"
          },
          "deliveryStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "SEARCHING_FOR_DRIVER", "MANUAL_INTERVENTION_REQUIRED", "ASSIGNED", "AT_RESTAURANT", "OUT_FOR_DELIVERY", "DELIVERED", "CANCELLED", "FAILED" ]
          },
          "paymentStatus" : {
            "type" : "string",
            "enum" : [ "INITIATED", "SUCCESS", "FAILED", "PENDING_COLLECTION", "COLLECTED", "PARTIALLY_REFUNDED", "REFUNDED", "REFUND_PENDING", "REFUND_FAILED" ]
          },
          "totalAmount" : {
            "type" : "number"
          },
          "itemTotal" : {
            "type" : "number"
          },
          "customerPlatformFee" : {
            "type" : "number"
          },
          "restaurantPlatformFee" : {
            "type" : "number"
          },
          "platformBonus" : {
            "type" : "number"
          },
          "restaurantDeliveryContribution" : {
            "type" : "number"
          },
          "restaurantPayout" : {
            "type" : "number"
          },
          "sgst" : {
            "type" : "number"
          },
          "cgst" : {
            "type" : "number"
          },
          "deliveryFee" : {
            "type" : "number"
          },
          "driverGrossPayout" : {
            "type" : "number"
          },
          "driverTaxes" : {
            "type" : "number"
          },
          "driverNetPayout" : {
            "type" : "number"
          },
          "quoteId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "rateBasePrice" : {
            "type" : "number"
          },
          "ratePerKm" : {
            "type" : "number"
          },
          "rateRestMaxContributionPercent" : {
            "type" : "number"
          },
          "rateFixedPlatformFee" : {
            "type" : "number"
          },
          "ratePlatformExcessCutPercent" : {
            "type" : "number"
          },
          "rateSgstPercent" : {
            "type" : "number"
          },
          "rateCgstPercent" : {
            "type" : "number"
          },
          "rateDeliverySgstPercent" : {
            "type" : "number"
          },
          "rateDeliveryCgstPercent" : {
            "type" : "number"
          },
          "distanceKm" : {
            "type" : "number"
          },
          "deliveryExecutiveId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "deliveryAddressId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "deliveryLat" : {
            "type" : "number",
            "format" : "double"
          },
          "deliveryLng" : {
            "type" : "number",
            "format" : "double"
          },
          "deliveryAddress" : {
            "type" : "string"
          },
          "pickupOtp" : {
            "type" : "string"
          },
          "otp" : {
            "type" : "string"
          },
          "estimatedPrepTimeMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "estimatedCompletionTime" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cancellationReason" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "deliveredAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          }
        }
      },
      "PageOrder" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Order"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "PageSupportTicket" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SupportTicket"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}